Moved logger function into separate package
