+++ tnr_bot/logger/__init__.py
@@ -1,0 +1,1 @@
+"""Custom logger with Telegram stream handler to send warning and errors into telegram logs chat"""
